updated to query for a particular track
Changes made to get one track alone with the ID given as an argument.

datasources - added a new method to call the rest endpoint to retrieve the track based on id.

schema - added a track in the query definition so that we can use it

resolver - add new field track to call the datasource and return the corresponding track back